reform distance trait
