+++ src/capsule.rs
@@ -1,6 +1,6 @@
 use mini_math::Point;
 
-use crate::{Distance, LineSegment};
+use crate::LineSegment;
 
 #[derive(Debug)]
 pub struct Capsule {
@@ -17,28 +17,5 @@ impl Capsule {
             axis: LineSegment::new(a, b),
             radius,
         }
-    }
-}
-
-impl Distance<Point> for Capsule {
-    /// Returns the distance between the sphere and a given point.
-    fn distance(&self, p: Point) -> f32 {
-        self.axis.distance(p) - self.radius
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_distance() {
-        let cap = Capsule::new(Point::new(0.0, 0.0, 0.0), Point::new(0.0, 5.0, 0.0), 1.0);
-
-        let p = Point::new(0.0, 0.0, -5.0);
-        assert_eq!(cap.distance(p), 4.0);
-
-        let p = Point::new(0.0, 10.0, 0.0);
-        assert_eq!(cap.distance(p), 4.0);
     }
 }
