+++ src/lib.rs
@@ -3,23 +3,23 @@
 mod capsule;
 mod closest_point;
 mod collision;
+mod distance;
 mod intersection;
 mod line;
 mod line_segment;
 mod plane;
 mod ray;
 mod sphere;
-mod traits;
 mod triangle;
 
 pub use capsule::*;
 pub use closest_point::*;
 pub use collision::*;
+pub use distance::*;
 pub use intersection::*;
 pub use line::*;
 pub use line_segment::*;
 pub use plane::*;
 pub use ray::*;
 pub use sphere::*;
-pub use traits::*;
 pub use triangle::*;
