+++ src/plane.rs
@@ -1,4 +1,4 @@
-use crate::{Distance, Triangle};
+use crate::Triangle;
 use mini_math::{Point, Vector3};
 
 /// An infinite plane.
@@ -30,59 +30,11 @@ impl Plane {
             d: Vector3::from(p).dot(normal),
         }
     }
-
-    /// Returns the point on the plane that is closest to the given point.
-    pub fn point_closest_to(&self, p: Point) -> Point {
-        let distance = self.distance(p);
-        p - self.normal * distance
-    }
 }
 
 impl From<&Triangle> for Plane {
     /// Convert a vector into a point
     fn from(t: &Triangle) -> Self {
         Plane::from_points(t.a, t.b, t.c)
-    }
-}
-
-impl Distance<Point> for Plane {
-    /// Returns the signed distance between the plane and a given point.
-    fn distance(&self, p: Point) -> f32 {
-        self.normal.dot(Vector3::from(p)) - self.d
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_distance() {
-        let plane = Plane::from_points(
-            Point::new(-1.0, 0.0, -1.0),
-            Point::new(1.0, 0.0, -1.0),
-            Point::new(0.0, 0.0, 1.0),
-        );
-
-        let p = Point::new(3.0, 1.0, 2.0);
-        assert_eq!(plane.distance(p), 1.0);
-
-        let p = Point::new(-2.0, -1.0, -3.0);
-        assert_eq!(plane.distance(p), -1.0);
-    }
-
-    #[test]
-    fn test_closest_point() {
-        let plane = Plane::from_points(
-            Point::new(-1.0, 0.0, -1.0),
-            Point::new(1.0, 0.0, -1.0),
-            Point::new(0.0, 0.0, 1.0),
-        );
-
-        let p = Point::new(2.0, 1.0, 3.0);
-        assert_eq!(plane.point_closest_to(p), Point::new(2.0, 0.0, 3.0));
-
-        let p = Point::new(-2.0, -1.0, -3.0);
-        assert_eq!(plane.point_closest_to(p), Point::new(-2.0, 0.0, -3.0));
     }
 }
