--- src/traits.rs
@@ -1,4 +1,0 @@
-pub trait Distance<Other> {
-    /// The distance between two objects
-    fn distance(&self, other: Other) -> f32;
-}
