+++ src/triangle.rs
@@ -59,34 +59,7 @@ impl Triangle {
         true
     }
 
-    /// Returns the point on the triangle that is closest to the given point.
-    pub fn point_closest_to(&self, p: Point) -> Point {
-        let plane = Plane::from(self);
-        let q = plane.point_closest_to(p);
-
-        let coordinates = self.barycentric_coordinates(q);
-        if coordinates.x >= 0.0 && coordinates.y >= 0.0 && coordinates.z >= 0.0 {
-            return q;
-        }
-
-        let p0 = Self::point_closest_to_edge(self.a, self.b, p);
-        let p1 = Self::point_closest_to_edge(self.b, self.c, p);
-        let p2 = Self::point_closest_to_edge(self.c, self.a, p);
-
-        let d0 = (p0 - p).magnitude_squared();
-        let d1 = (p1 - p).magnitude_squared();
-        let d2 = (p2 - p).magnitude_squared();
-
-        if d0 < d1 && d0 < d2 {
-            p0
-        } else if d1 < d0 && d1 < d2 {
-            p1
-        } else {
-            p2
-        }
-    }
-
-    fn point_closest_to_edge(e0: Point, e1: Point, p: Point) -> Point {
+    pub(crate) fn point_closest_to_edge(e0: Point, e1: Point, p: Point) -> Point {
         let edge = e1 - e0;
         let edge_length = edge.magnitude();
         let edge_direction = edge / edge_length;
@@ -99,28 +72,5 @@ impl Triangle {
         } else {
             e0 + edge_direction * d
         }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_closest_point() {
-        let triangle = Triangle::new(
-            Point::new(-1.0, 0.0, -1.0),
-            Point::new(1.0, 0.0, -1.0),
-            Point::new(0.0, 0.0, 1.0),
-        );
-
-        let p = Point::new(0.0, 1.0, 0.0);
-        assert_eq!(triangle.point_closest_to(p), Point::new(0.0, 0.0, 0.0));
-
-        let p = Point::new(0.0, 1.0, 2.0);
-        assert_eq!(triangle.point_closest_to(p), Point::new(0.0, 0.0, 1.0));
-
-        let p = Point::new(0.0, -1.0, -2.0);
-        assert_eq!(triangle.point_closest_to(p), Point::new(0.0, 0.0, -1.0));
     }
 }
